Get antineoplastic drugs and drug regimens

Downloads preprocessed datasets to a local cache directory and returns a selected set of drugs based on various criteria set by the user.

The dataset comes as a list object, with two elements:

metadata - a data frame with metadata regarding drug resources used

records - a data frame with drug records

Usage

get_drugs(
  cache_dir = NA,
  force_download = F,
  exclude_salt_forms = T,
  exclude_adc = T,
  drug_is_targeted = F,
  drug_is_approved = F,
  drug_target = NULL,
  drug_action_type = NULL,
  drug_indication_main = NULL,
  source_opentargets_only = F,
  drug_has_blackbox_warning = F,
  drug_approved_later_than = 1939,
  drug_minimum_phase_any_indication = 0,
  output_resolution = "drug2target2indication",
  inhibitor_only = F,
  is_alkylating_agent = F,
  is_angiogenesis_inhibitor = F,
  is_anthracycline = F,
  is_antimetabolite = F,
  is_ar_antagonist = F,
  is_bet_inhibitor = F,
  is_hedgehog_antagonist = F,
  is_hdac_inhibitor = F,
  is_hormone_therapy = F,
  is_iap_inhibitor = F,
  is_immune_checkpoint_inhibitor = F,
  is_kinase_inhibitor = F,
  is_monoclonal_antibody = F,
  is_parp_inhibitor = F,
  is_platinum_compound = F,
  is_proteasome_inhibitor = F,
  is_topoisomerase_inhibitor = F,
  is_tubulin_inhibitor = F
)

Arguments

cache_dir

local cache directory for data retrieval

force_download

force download data from remote repository even if data exists in cache

exclude_salt_forms

exclude salt forms of drugs

exclude_adc

exclude antibody-drug conjugates (ADCs)

drug_is_targeted

logical indicating if resulting drug records should contain drugs with known drug targets only

drug_is_approved

logical indicating if resulting drug records should contain approved drugs only

drug_target

character vector with drug targets (gene symbols) for drug records included in results

drug_action_type

character vector with drug action types to include in drug record list - possible values "INHIBITOR","AGONIST","MODULATOR","ANTAGONIST","BLOCKER","ACTIVATOR","BINDING AGENT","OPENER", "STABILISER","CROSS-LINKING AGENT",DISRUPTING AGENT","OTHER"

drug_indication_main

character vector with main tumor types for which drug(s) are indicated possible values "Adrenal Gland","Biliary Tract","Bladder/Urinary Tract","Bone","Breast","Cervix","CNS/Brain", "Colon/Rectum","Esophagus/Stomach","Eye","Head and Neck","Kidney","Liver","Lung","Lymphoid","Myeloid","Ovary/Fallopian Tube", "Pancreas","Penis","Peripheral Nervous System","Peritoneum","Pleura","Prostate","Skin","Soft Tissue","Testis","Thymus", "Thyroid","Uterus","Vulva/Vagina"

source_opentargets_only

logical indicating if resulting drug records should contain drug records from Open Targets Platform only

drug_has_blackbox_warning

logical indicating if resulting drug records should contain drugs with black box warnings only

drug_approved_later_than

only include records for drugs approved later than this date (year)

drug_minimum_phase_any_indication

only include drug records that are in a clinical phase (any indication) greater or equal than this phase

output_resolution

dictate output record resolution ('drug','drug2target','drug2target2indication')

inhibitor_only

logical indicating to return only drugs with inhibitory mechanism-of-action

is_alkylating_agent

logical indicating if only this drug class is to be considered

is_angiogenesis_inhibitor

logical indicating if only this drug class is to be considered

is_anthracycline

logical indicating if only this drug class is to be considered

is_antimetabolite

logical indicating if only this drug class is to be considered

is_ar_antagonist

logical indicating if only this drug class is to be considered

is_bet_inhibitor

logical indicating if only this drug class is to be considered

is_hedgehog_antagonist

logical indicating if only this drug class is to be considered

is_hdac_inhibitor

logical indicating if only this drug class is to be considered

is_hormone_therapy

logical indicating if only this drug class is to be considered

is_iap_inhibitor

logical indicating if only this drug class is to be considered

is_immune_checkpoint_inhibitor

logical indicating if only this drug class is to be considered

is_kinase_inhibitor

logical indicating if only this drug class is to be considered

is_monoclonal_antibody

logical indicating if only this drug class is to be considered

is_parp_inhibitor

logical indicating if only this drug class is to be considered

is_platinum_compound

logical indicating if only this drug class is to be considered

is_proteasome_inhibitor

logical indicating if only this drug class is to be considered

is_topoisomerase_inhibitor

logical indicating if only this drug class is to be considered

is_tubulin_inhibitor

logical indicating if only this drug class is to be considered

Value

The records data frame contains the following columns (only selected columns will be shown based on the value of output_resolution)

drug_id - drug identifier (pharmaOncoX)

drug_name - primary drug name (upper case, NCI Thesaurus)

drug_type - type of drug molecule (Antibody, small molecule etc)

molecule_chembl_id - ChEMBL compound identifier

drug_action_type - main action elicited by drug (antagonist, inhibitor, stabiliser etc)

drug_alias - collection of unambiguous drug aliases (separated by '|')

nci_concept_definition - detailed description of drug mechanism-of-action (NCI Thesaurus)

opentargets - logical - drug is found in the Open Targets Platform resource

is_salt - logical - drug record represents a salt form (excluded by default)

is_adc - logical - drug record represents an antibody-drug conjugate (ADC - excluded by default)

drug_blacbox_warning - logical indicating if drug has blackbox warning

nci_t - NCI thesaurus identifier

immune_checkpoint_inhibitor - logical indicating if drug is an immune checkpoint inhibitor

topoisomerase_inhibitor - logical indicating if drug is a topoisomerase inhibitor

iap_inhibitor - logical indicating if drug is an IAP inhibitor

tubulin_inhibitor - logical indicating if drug is a tubulin inhibitor

kinase_inhibitor - logical indicating if drug is a kinase inhibitor

hdac_inhibitor - logical indicating if drug is a HDAC inhibitor

parp_inhibitor - logical indicating if drug is a PARP inhibitor

bet_inhibitor - logical indicating if drug is a BET inhibitor

ar_antagonist - logical indicating if drug is an androgen receptor antagonist

monoclonal_antibody - logical indicating if drug is an antibody

antimetabolite - logical indicating if drug is an antimetabolite

angiogenesis_inhibitor - logical indicating if drug is an angiogenesis inhibitor

alkylating_agent - logical indicating if drug is an alkylating agent

platinum_compound - logical indicating if drug is a platinum compound

anthracycline - logical indicating if drug is an anthracycline

proteasome_inhibitor - logical indicating if drug is a proteasome inhibitor

hormone_therapy - logical indicating if drug is a hormone therapy

hedgehog_antagonist - logical indicating if drug is a hedgehog pathway antagonist

fda_epc_category - FDA Established Pharmacologic Class (EPC) text phrase

target_symbol - gene symbol of drug target

target_entrezgene - Entrez gene identifier of drug target

target_genename - gene name/description of drug target

target_ensembl_gene_id - Ensembl gene identifier of drug target

target_type - type of drug target (single protein, protein family etc.)

drug_max_phase_indication - maximum clinical phase for drug (given indication)

drug_approved_indication - logical indicating if drug has an approved indication

drug_year_first_approval - year drug was first approved

drug_max_ct_phase - maximum clinical phase for drug (any indication)

disease_efo_id - EFO (Experimental Factor Ontology) identifier for drug indication

disease_efo_label - EFO (Experimental Factor Ontology) label for drug indication

primary_site - primary tumor site/type (obtained through https://github.com/sigven/oncoPhenoMap)

drug_clinical_source - underlying source for drug entry (DailyMed, clinicaltrials.gov, FDA etc.)

drug_clinical_id - drug clinical identifier (clinicaltrials.gov, DailyMed, FDA etc.)